bugfix: 修复数据集同版本重试静默复用旧 release，可能导致发布任务假成功与状态卡死的问题
